requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.4
numpy==2.2.6
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
whitenoise==6.9.0
